first

+++ Day2Codehere/Practice/Practice1.cpp
@@ -6,7 +6,7 @@ int main() {
     int n, m;
     cin >> n >> m;
     
-    vector<int> a(n), b(m);
+    vector<long long> a(n), b(m);
     for (int i = 0; i < n; i++) cin >> a[i];
     for (int i = 0; i < m; i++) cin >> b[i];
